# app/controllers/auth.go
package controllers

import (
	"time"

	"github.com/AlvinMrema/kighala-api/app/models"
	"github.com/AlvinMrema/kighala-api/pkg/utils"
	jwt_utils "github.com/AlvinMrema/kighala-api/pkg/utils/jwt_utils"
	dbConfig "github.com/AlvinMrema/kighala-api/platform/database"
	database "github.com/AlvinMrema/kighala-api/platform/database/sqlc"
	"github.com/gofiber/fiber/v2"
	"github.com/golang-jwt/jwt/v5"

	"github.com/google/uuid"
	"golang.org/x/crypto/bcrypt"
)

// SignUp func for user signin.
// @Description User signup.
// @Summary user signup
// @Tags Auth
// @Accept json
// @Produce json
// @Param user body models.AuthCredentials true "Username and Password"
// @Success 201 {object} models.User
// @Router /auth/signup [post]
func SignUp(c *fiber.Ctx) error {
	data := models.User{}
	if err := c.BodyParser(&data); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	hashPassword, err := bcrypt.GenerateFromPassword([]byte(data.Password), bcrypt.DefaultCost)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	role := utils.User.String()

	// Create database connection.
	db, err := dbConfig.OpenDBConnection()
	if err != nil {
		// Return status 500 and database connection error.
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	// Creating the 'user' to database
	dbResult, err := db.DB.CreateUser(c.Context(), database.CreateUserParams{
		ID:        uuid.New(),
		CreatedAt: time.Now().UTC(),
		UpdatedAt: time.Now().UTC(),
		Username:  data.Username,
		Password:  string(hashPassword),
		Role:      role,
	})
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	return c.Status(fiber.StatusCreated).JSON(fiber.Map{
		"error":   false,
		"results": dbResult,
	})
}

// SignIn func for user signin.
// @Description User signin.
// @Summary user signin
// @Tags Auth
// @Accept json
// @Produce json
// @Param user body models.AuthCredentials true "Username and Password"
// @Success 201 {object} models.User
// @Router /auth/signin [post]
func SignIn(c *fiber.Ctx) error {
	data := models.User{}
	if err := c.BodyParser(&data); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	// Create database connection.
	db, err := dbConfig.OpenDBConnection()
	if err != nil {
		// Return status 500 and database connection error.
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	dbResult, err := db.DB.GetUserByUsername(c.Context(), data.Username)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	err = bcrypt.CompareHashAndPassword([]byte(dbResult.Password), []byte(data.Password))
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	expirationTime := time.Now().Add(10 * time.Minute)

	claims := &jwt_utils.Claims{
		Username: dbResult.Username,
		Role:     dbResult.Role,
		RegisteredClaims: jwt.RegisteredClaims{
			ExpiresAt: jwt.NewNumericDate(expirationTime),
		},
	}

	jwtOutput, err := jwt_utils.JWTGenerateToken(claims, expirationTime)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	return c.Status(fiber.StatusCreated).JSON(fiber.Map{
		"error":   false,
		"results": jwtOutput,
	})
}